+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * New players show the full interface by default
 * Mouse hiding now works on a timer
+* Auto complete separates words
 
 ### Fixed
 * Possible crash trying to hide the mouse pointer